Backend/requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
firebase-admin==5.2.0
gunicorn==20.1.0
Pyrebase4==4.5.0
requests==2.27.1
dj-database-url==0.5.0